新建分支，准备决战ES分析

+++ Apps/EverStationII/Magic_ES/ESMagicView.cpp
@@ -498,7 +498,7 @@ void CESMagicView::OnBtnPreview()
 	{
 		CComboBox* pCmbCtrl = (CComboBox*)GetDlgItem(IDC_CMB_ES_TYPE);
 		int nSel = pCmbCtrl->GetCurSel();
-		int nStreamType = pCmbCtrl->GetItemData(nSel);
+		int nStreamType = (int)pCmbCtrl->GetItemData(nSel);
 
 		//switch (nStreamType & (~STREAM_FILE))
 		//{
